backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
python-multipart==0.0.6
sqlalchemy==2.0.23
alembic==1.12.1
python-iptables==1.0.1
dnspython==2.4.2
requests==2.31.0
schedule==1.2.0
aiocron==1.8
aiohttp==3.9.1
pydantic==2.5.0
pydantic-settings==2.1.0
ipaddress==1.0.23
netaddr==0.10.1
aiofiles==23.2.1
python-dateutil==2.8.2
bcrypt==4.1.2 